Refactored to make ObjectManager the Observer. ChasingElement now move in relation to Player even if Player stops.
